pyproject.toml
[tool.poetry]
name = "npmview"
version = "0.1.0"
description = ""
authors = ["Taiwo Ajiboye <boldrackdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
